+++ examples/src/3d/cube.rs
@@ -3,51 +3,47 @@ use crayon_testbed::prelude::*;
 
 struct Window {
     world: World<SimpleRenderer>,
-    room: Entity,
     canvas: ConsoleCanvas,
+    cube: Entity,
 }
 
 impl Window {
     fn new(engine: &mut Engine) -> Result<Self> {
         //
-        let world_resources = WorldResources::new(engine)?;
+        let world_resources = (WorldResources::new(engine)?).shared();
+
         let ctx = engine.context();
-        let pipeline = SimpleRenderer::new(&ctx, world_resources.shared())?;
-        let mut world = World::new(world_resources.shared(), pipeline);
+        let pipeline = SimpleRenderer::new(&ctx, world_resources.clone())?;
+
+        let mut world = World::new(world_resources.clone(), pipeline);
 
         //
-        let prefab = ctx.res.load("res:cornell_box.obj")?;
-        ctx.res.wait(prefab)?;
-        let room = world.instantiate(prefab).unwrap();
+        let cube = world.create();
 
-        // Lets give shortBox some red color.
+        let mesh = world_resources.meshes.cube;
+        world.renderables.add_mesh(cube, mesh);
+
+        // Lets give cube a crate texture.
         let mut m = SimpleMaterial::default();
-        let e = world.find("cornell_box.obj/shortBox").unwrap();
-        m.diffuse = [255, 100, 100, 255].into();
-        world.renderer.add(e, m);
-
-        let e = world.find("cornell_box.obj/tallBox").unwrap();
-        m.diffuse = [55, 55, 255, 255].into();
-        world.renderer.add(e, m);
+        m.diffuse_texture = Some(ctx.res.load("res:crate.bmp")?);
+        world.renderer.add(cube, m);
 
         //
         let lit = world.create();
-        let rotation = math::Euler::new(math::Deg(45.0), math::Deg(0.0), math::Deg(0.0));
         world.renderables.add_lit(lit, Lit::default());
-        world.scene.set_rotation(lit, rotation);
 
         //
         let camera = world.create();
-        let params = Camera::ortho(4.8, 3.2, 0.1, 5.0);
+        let params = Camera::ortho(3.2, 2.4, 0.1, 5.0);
         let center = [0.0, 0.0, 0.0];
         world.renderables.add_camera(camera, params);
-        world.scene.set_position(camera, [0.0, 2.0, -2.0]);
+        world.scene.set_position(camera, [0.0, 0.0, -2.0]);
         world.scene.look_at(camera, center, [0.0, 1.0, 0.0]);
 
         Ok(Window {
-            room: room,
             world: world,
             canvas: ConsoleCanvas::new(&ctx, None)?,
+            cube: cube,
         })
     }
 }
@@ -63,7 +59,7 @@ impl Application for Window {
                 math::Deg(0.0),
             );
 
-            self.world.scene.rotate(self.room, rotation);
+            self.world.scene.rotate(self.cube, rotation);
         }
 
         self.canvas.render(ctx);
@@ -77,7 +73,7 @@ impl Application for Window {
 }
 
 fn main() {
-    let params = crayon_testbed::settings("CR: Prefab", (640, 480));
+    let params = crayon_testbed::settings("CR: Cube", (640, 480));
     let mut engine = Engine::new_with(&params).unwrap();
     let res = crayon_testbed::find_res_dir();
     engine.res.mount("res", res).unwrap();