+++ src/video/backends/gl/types.rs
@@ -12,6 +12,7 @@ where
     T: Sized + Clone,
 {
     pub buf: Vec<Option<T>>,
+    pub versions: Vec<u32>,
 }
 
 impl<T> DataVec<T>
@@ -19,25 +20,38 @@ where
     T: Sized + Clone,
 {
     pub fn new() -> Self {
-        DataVec { buf: Vec::new() }
+        DataVec {
+            buf: Vec::new(),
+            versions: Vec::new(),
+        }
     }
 
     pub fn get<H>(&self, handle: H) -> Option<&T>
     where
         H: Borrow<handle::Handle>,
     {
-        self.buf
-            .get(handle.borrow().index() as usize)
-            .and_then(|v| v.as_ref())
+        let index = handle.borrow().index() as usize;
+        if let Some(&v) = self.versions.get(index) {
+            if v == handle.borrow().version() {
+                return self.buf[index].as_ref();
+            }
+        }
+
+        None
     }
 
     pub fn get_mut<H>(&mut self, handle: H) -> Option<&mut T>
     where
         H: Borrow<handle::Handle>,
     {
-        self.buf
-            .get_mut(handle.borrow().index() as usize)
-            .and_then(|v| v.as_mut())
+        let index = handle.borrow().index() as usize;
+        if let Some(&v) = self.versions.get(index) {
+            if v == handle.borrow().version() {
+                return self.buf[index].as_mut();
+            }
+        }
+
+        None
     }
 
     pub fn create<H>(&mut self, handle: H, value: T)
@@ -49,9 +63,11 @@ where
 
         if self.buf.len() <= index {
             self.buf.resize(index + 1, None);
+            self.versions.resize(index + 1, 1);
         }
 
         self.buf[index] = Some(value);
+        self.versions[index] = handle.version();
     }
 
     pub fn free<H>(&mut self, handle: H) -> Option<T>