+++ CHANGELOG.md
@@ -9,7 +9,13 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 
 ## [Unreleased]
 
-## 0.5.0 - 2018-08-13
-* Rebases the initial release from v0.5.0.
+## [0.5.1] - 2018-09-01
+* Re-export macros from cgmath.
+* Add optional attribute field in shader.
+* Fix un-expected side-effects of GLVisitor::bind_texture.
 
-[Unreleased]: https://github.com/shawnscode/crayon/compare/v0.5.0...HEAD
+## 0.5.0 - 2018-08-13
+* Rebase the initial release from v0.5.0.
+
+[Unreleased]: https://github.com/shawnscode/crayon/compare/v0.5.1...HEAD
+[0.5.1]: https://github.com/shawnscode/crayon/compare/v0.5.0...v0.5.1